move homedir logic from client-go into kubeconfig package
